requirements.txt
# Build dependencies
Flask==1.1.1
Flask-API==1.1
Flask-SQLAlchemy==2.4.1
psycopg2-binary==2.8.3
python-dotenv==0.10.3
redis>=2.10

# Runtime
gunicorn==19.9.0
honcho==1.0.1

# Testing
nose==1.3.7
pinocchio==0.4.2
httpie>=1.0.3
pylint>=2.4.1
factory-boy==2.12.0
# Code coverage
coverage==4.5.4
codecov==2.0.15
